1. 멀티모듈에서 일반 프로젝트로 구조 변경 2. 전체적인 패키지 변경. 3. 입차 전/후방 1차개발 완료 - 세부적인 내용은 테스트하면서 변경 4. Display api 생성. 5. Display Message Format 상수 설정. 6. Display Message 생성 로직 개발 7 netty multi client 생성.
