--- breaker/build.gradle
@@ -1,13 +1,0 @@
-dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
-    implementation 'org.springframework.boot:spring-boot-starter-aop'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.retry:spring-retry'
-    implementation 'org.modelmapper:modelmapper:2.3.0'
-    implementation 'commons-io:commons-io:2.6' // image file encode
-
-
-
-//    compile group: 'io.netty', name: 'netty-all', version: '4.1.25.Final'
-}
