--- breaker/src/main/java/kr/co/glnt/config/ServerConfig.java
@@ -1,25 +1,0 @@
-package kr.co.glnt.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.DefaultUriBuilderFactory;
-
-@Setter @Getter
-@Configuration
-@ConfigurationProperties("spring.server-config")
-public class ServerConfig {
-    private String gpmsUrl;
-    private String ngisUrl;
-    private String serverName;
-    private String fileErrPath;
-}
