--- breaker/src/main/resources/application.yml
@@ -1,13 +1,0 @@
-server:
-  port: 9999
-
-spring:
-  profiles:
-    active: GATESVR1
-  server-config:
-    server-name: GATESVR1
-    gpms-url: "http://ec2-54-180-8-134.ap-northeast-2.compute.amazonaws.com:7000"
-    ngis-url: "http://localhost:7777"
-    file-err-in-path: "C:\\park\\in-error"
-    file-err-out-path: "C:\\park\\out-error"
----
