--- src/main/java/kr/co/glnt/modules/ModulesApplication.java
@@ -1,13 +1,0 @@
-package kr.co.glnt.modules;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ModulesApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ModulesApplication.class, args);
-    }
-
-}
