+++ src/main/java/kr/co/glnt/relay/run/HealthCheckRunner.java
@@ -1,0 +1,4 @@
+package kr.co.glnt.relay.run;
+
+public class HealthCheckRunner {
+}
